Core/Misc: Fixed nopch MSVC build
(cherry picked from commit e977bb4c4de3608cbbfe0b2d28cd0f4176a44777)
