+++ src/server/game/Handlers/MovementHandler.cpp
@@ -19,6 +19,7 @@
 #include "Battleground.h"
 #include "Common.h"
 #include "Corpse.h"
+#include "FlightPathMovementGenerator.h"
 #include "GameTime.h"
 #include "Garrison.h"
 #include "InstancePackets.h"
@@ -34,7 +35,6 @@
 #include "MovementGenerator.h"
 #include "Transport.h"
 #include "Vehicle.h"
-#include "WaypointMovementGenerator.h"
 #include "SpellMgr.h"
 
 #define MOVEMENT_PACKET_TIME_DELAY 0
